Added CI (basic and advanced)
Also added scalafmtcheck - This is part of CI but can be changed if we wish so.

in sbt terminal run scalafmtCheck to find if any files (and which) are not formatted as we like.

then run scalafmt to automatically reformat the files. I believe this can be targeted at single files as well but have not tried this.